update: create custom hook to handle language changes
